requirements.txt
keyboard==0.13.5
PyYAML==5.4.1
selenium==3.141.0
urllib3==1.26.4